refactor!: Reorganize exports and modules exposed
