+++ workspaces/src/lib.rs
@@ -3,16 +3,18 @@ mod cargo;
 #[cfg(feature = "unstable")]
 pub use cargo::compile_project;
 
-mod network;
 mod rpc;
 mod types;
 mod worker;
 
+pub mod network;
 pub mod prelude;
+pub mod result;
 
-pub use network::result;
-pub use network::transaction::Function;
-pub use network::{Account, AccountDetails, Block, Contract, DevNetwork, Network};
+pub use network::variants::{DevNetwork, Network};
+pub use types::account::{Account, AccountDetails, Contract};
+pub use types::block::Block;
+pub use types::transaction::Function;
 pub use types::{AccessKey, AccountId, BlockHeight, CryptoHash, InMemorySigner};
 pub use worker::{
     mainnet, mainnet_archival, sandbox, testnet, testnet_archival, with_mainnet,
