+++ workspaces/src/network/mod.rs
@@ -1,154 +1,16 @@
-mod account;
-mod block;
 mod info;
 mod mainnet;
-pub mod result;
 mod sandbox;
 mod server;
 mod testnet;
+pub(crate) mod variants;
 
-pub mod transaction;
+pub(crate) use variants::DEV_ACCOUNT_SEED;
 
-use async_trait::async_trait;
-
-use near_jsonrpc_client::methods::sandbox_patch_state::RpcSandboxPatchStateRequest;
-use near_primitives::state_record::StateRecord;
-
-pub(crate) use crate::network::info::Info;
-use crate::rpc::client::Client;
-use crate::rpc::patch::ImportContractTransaction;
-use crate::types::{AccountId, KeyType, SecretKey};
-use crate::Worker;
-
-pub use crate::network::account::{Account, AccountDetails, Contract};
-pub use crate::network::block::Block;
-pub use crate::network::mainnet::Mainnet;
-pub use crate::network::result::{CallExecution, CallExecutionDetails, ViewResultDetails};
-pub use crate::network::sandbox::Sandbox;
-pub use crate::network::testnet::Testnet;
-
-pub(crate) const DEV_ACCOUNT_SEED: &str = "testificate";
-
-pub trait NetworkClient {
-    fn client(&self) -> &Client;
-}
-
-pub trait NetworkInfo {
-    fn info(&self) -> &Info;
-}
-
-#[async_trait]
-pub trait TopLevelAccountCreator {
-    async fn create_tla(
-        &self,
-        id: AccountId,
-        sk: SecretKey,
-    ) -> anyhow::Result<CallExecution<Account>>;
-
-    async fn create_tla_and_deploy(
-        &self,
-        id: AccountId,
-        sk: SecretKey,
-        wasm: &[u8],
-    ) -> anyhow::Result<CallExecution<Contract>>;
-}
-
-// NOTE: Not all networks/runtimes will have the ability to be able to do dev_deploy.
-// This trait acts as segmented boundary for only specific networks such as sandbox and testnet.
-pub trait AllowDevAccountCreation {}
-
-#[async_trait]
-pub trait DevAccountDeployer {
-    async fn dev_generate(&self) -> (AccountId, SecretKey);
-    async fn dev_create_account(&self) -> anyhow::Result<Account>;
-    async fn dev_deploy(&self, wasm: &[u8]) -> anyhow::Result<Contract>;
-}
-
-#[async_trait]
-impl<T> DevAccountDeployer for T
-where
-    T: TopLevelAccountCreator + NetworkInfo + AllowDevAccountCreation + Send + Sync,
-{
-    async fn dev_generate(&self) -> (AccountId, SecretKey) {
-        let id = crate::rpc::tool::random_account_id();
-        let sk = SecretKey::from_seed(KeyType::ED25519, DEV_ACCOUNT_SEED);
-        (id, sk)
-    }
-
-    async fn dev_create_account(&self) -> anyhow::Result<Account> {
-        let (id, sk) = self.dev_generate().await;
-        let account = self.create_tla(id.clone(), sk).await?;
-        account.into()
-    }
-
-    async fn dev_deploy(&self, wasm: &[u8]) -> anyhow::Result<Contract> {
-        let (id, sk) = self.dev_generate().await;
-        let contract = self.create_tla_and_deploy(id.clone(), sk, wasm).await?;
-        contract.into()
-    }
-}
-
-pub trait AllowStatePatching {}
-
-#[async_trait]
-pub trait StatePatcher {
-    async fn patch_state(
-        &self,
-        contract_id: &AccountId,
-        key: &[u8],
-        value: &[u8],
-    ) -> anyhow::Result<()>;
-
-    fn import_contract<'a, 'b>(
-        &'b self,
-        id: &AccountId,
-        worker: &'a Worker<impl Network>,
-    ) -> ImportContractTransaction<'a, 'b>;
-}
-
-#[async_trait]
-impl<T> StatePatcher for T
-where
-    T: AllowStatePatching + NetworkClient + Send + Sync,
-{
-    async fn patch_state(
-        &self,
-        contract_id: &AccountId,
-        key: &[u8],
-        value: &[u8],
-    ) -> anyhow::Result<()> {
-        let state = StateRecord::Data {
-            account_id: contract_id.to_owned(),
-            data_key: key.to_vec(),
-            value: value.to_vec(),
-        };
-        let records = vec![state];
-
-        // NOTE: RpcSandboxPatchStateResponse is an empty struct with no fields, so don't do anything with it:
-        let _patch_resp = self
-            .client()
-            .query(&RpcSandboxPatchStateRequest { records })
-            .await
-            .map_err(|err| anyhow::anyhow!("Failed to patch state: {:?}", err))?;
-
-        Ok(())
-    }
-
-    fn import_contract<'a, 'b>(
-        &'b self,
-        id: &AccountId,
-        worker: &'a Worker<impl Network>,
-    ) -> ImportContractTransaction<'a, 'b> {
-        ImportContractTransaction::new(id.to_owned(), worker.client(), self.client())
-    }
-}
-
-pub trait Network: TopLevelAccountCreator + NetworkInfo + NetworkClient + Send + Sync {}
-
-impl<T> Network for T where T: TopLevelAccountCreator + NetworkInfo + NetworkClient + Send + Sync {}
-
-/// DevNetwork is a Network that can call into `dev_create` and `dev_deploy` to create developer accounts.
-pub trait DevNetwork: AllowDevAccountCreation + Network {}
-
-// Implemented by default if we have `AllowDevAccountCreation`
-impl<T> DevNetwork for T where T: AllowDevAccountCreation + Network {}
+pub use self::info::Info;
+pub use self::mainnet::Mainnet;
+pub use self::sandbox::Sandbox;
+pub use self::testnet::Testnet;
+pub use self::variants::{
+    AllowDevAccountCreation, DevAccountDeployer, NetworkClient, NetworkInfo, TopLevelAccountCreator,
+};
