+++ workspaces/src/network/testnet.rs
@@ -7,14 +7,11 @@ use url::Url;
 use near_primitives::views::{ExecutionStatusView, FinalExecutionStatus};
 
 use crate::network::Info;
-use crate::network::{
-    Account, AllowDevAccountCreation, CallExecution, CallExecutionDetails, NetworkClient,
-    NetworkInfo, TopLevelAccountCreator,
-};
-use crate::result::ExecutionOutcome;
+use crate::network::{AllowDevAccountCreation, NetworkClient, NetworkInfo, TopLevelAccountCreator};
+use crate::result::{CallExecution, CallExecutionDetails, ExecutionOutcome};
 use crate::rpc::{client::Client, tool};
 use crate::types::{AccountId, InMemorySigner, SecretKey};
-use crate::{Contract, CryptoHash};
+use crate::{Account, Contract, CryptoHash};
 
 const RPC_URL: &str = "https://rpc.testnet.near.org";
 const HELPER_URL: &str = "https://helper.testnet.near.org";
