+++ workspaces/src/types/mod.rs
@@ -1,3 +1,7 @@
+pub(crate) mod account;
+pub(crate) mod block;
+pub(crate) mod transaction;
+
 /// Types copied over from near_primitives since those APIs are not yet stable.
 /// and internal libraries like near-jsonrpc-client requires specific versions
 /// of these types which shouldn't be exposed either.
