+++ workspaces/src/types/transaction.rs
@@ -7,13 +7,14 @@ use near_primitives::transaction::{
 };
 use near_primitives::views::FinalExecutionOutcomeView;
 
-use crate::network::{CallExecution, CallExecutionDetails, Network, ViewResultDetails};
+use crate::result::{CallExecution, CallExecutionDetails, ViewResultDetails};
 use crate::rpc::client::{
     send_batch_tx_and_retry, Client, DEFAULT_CALL_DEPOSIT, DEFAULT_CALL_FN_GAS,
 };
 use crate::types::{AccessKey, AccountId, Balance, Gas, InMemorySigner, PublicKey, SecretKey};
 use crate::worker::Worker;
 use crate::Account;
+use crate::Network;
 
 const MAX_GAS: Gas = 300_000_000_000_000;
