+++ workspaces/src/worker/impls.rs
@@ -1,17 +1,18 @@
 use std::collections::HashMap;
 
 use async_trait::async_trait;
+use near_jsonrpc_client::methods::sandbox_patch_state::RpcSandboxPatchStateRequest;
+use near_primitives::state_record::StateRecord;
 use near_primitives::types::{Balance, StoreKey};
 
-use crate::network::{
-    Account, AllowDevAccountCreation, Block, CallExecution, CallExecutionDetails, Contract,
-    NetworkClient, NetworkInfo, StatePatcher, TopLevelAccountCreator, ViewResultDetails,
-};
+use crate::network::{AllowDevAccountCreation, NetworkClient, NetworkInfo, TopLevelAccountCreator};
 use crate::network::{Info, Sandbox};
+use crate::result::{CallExecution, CallExecutionDetails, ViewResultDetails};
 use crate::rpc::client::{Client, DEFAULT_CALL_DEPOSIT, DEFAULT_CALL_FN_GAS};
 use crate::rpc::patch::ImportContractTransaction;
 use crate::types::{AccountId, Gas, InMemorySigner, SecretKey};
 use crate::worker::Worker;
+use crate::{Account, Block, Contract};
 use crate::{AccountDetails, Network};
 
 impl<T> Clone for Worker<T> {
@@ -56,26 +57,36 @@ where
     }
 }
 
-#[async_trait]
-impl<T> StatePatcher for Worker<T>
-where
-    T: StatePatcher + Send + Sync,
-{
-    async fn patch_state(
+impl Worker<Sandbox> {
+    pub async fn patch_state(
         &self,
         contract_id: &AccountId,
         key: &[u8],
         value: &[u8],
     ) -> anyhow::Result<()> {
-        self.workspace.patch_state(contract_id, key, value).await
+        let state = StateRecord::Data {
+            account_id: contract_id.to_owned(),
+            data_key: key.to_vec(),
+            value: value.to_vec(),
+        };
+        let records = vec![state];
+
+        // NOTE: RpcSandboxPatchStateResponse is an empty struct with no fields, so don't do anything with it:
+        let _patch_resp = self
+            .client()
+            .query(&RpcSandboxPatchStateRequest { records })
+            .await
+            .map_err(|err| anyhow::anyhow!("Failed to patch state: {:?}", err))?;
+
+        Ok(())
     }
 
-    fn import_contract<'a, 'b>(
+    pub fn import_contract<'a, 'b>(
         &'b self,
         id: &AccountId,
         worker: &'a Worker<impl Network>,
     ) -> ImportContractTransaction<'a, 'b> {
-        self.workspace.import_contract(id, worker)
+        ImportContractTransaction::new(id.to_owned(), worker.client(), self.client())
     }
 }
